# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GrabGo API
  description: |
    Backend API for GrabGo food delivery platform.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## User Roles
    - **customer**: Regular users who order food
    - **restaurant**: Restaurant owners/managers
    - **rider**: Delivery riders
    - **admin**: Platform administrators
  version: 1.0.0
  contact:
    name: GrabGo Support
  license:
    name: ISC
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.grabgo.com/api
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login/register endpoints
  schemas:
...
    ReferralCode:
      type: object
      properties:
        code:
          type: string
          example: "ABC12345"
        shareUrl:
          type: string
          example: "https://grabgo.app/r/ABC12345"
        totalReferrals:
          type: integer
        completedReferrals:
          type: integer
        pendingReferrals:
          type: integer
        totalEarned:
          type: number
        availableCredit:
          type: string
    Referral:
      type: object
      properties:
        id:
          type: string
        refereeName:
          type: string
        status:
          type: string
          enum: [pending_order, completed, expired]
        creditEarned:
          type: number
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    UserCredit:
      type: object
      properties:
        id:
          type: string
        amount:
          type: number
        source:
          type: string
          enum: [referral_received, referral_earned, promo, milestone_bonus]
        expiresAt:
          type: string
          format: date-time
        description:
          type: string
    PromotionalBanner:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        subtitle:
          type: string
        imageUrl:
          type: string
          format: uri
        targetUrl:
          type: string
        targetScreen:
          type: string
        priority:
          type: integer
        isActive:
          type: boolean
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /referral/my-code:
    get:
      tags: [Referrals]
      summary: Get user's referral code and stats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Referral code retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferralCode'
  /referral/my-referrals:
    get:
      tags: [Referrals]
      summary: Get list of user's referrals
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Referrals retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  referrals:
                    type: array
                    items:
                      $ref: '#/components/schemas/Referral'
  /referral/validate:
    post:
      tags: [Referrals]
      summary: Validate a referral code
      description: Public endpoint to validate referral codes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  example: "ABC12345"
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  referrerName:
                    type: string
                  discount:
                    type: number
                  minOrderValue:
                    type: number
                  validDays:
                    type: integer
                  isSystemCode:
                    type: boolean
  /referral/apply:
    post:
      tags: [Referrals]
      summary: Apply referral code to new user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                deviceId:
                  type: string
      responses:
        '200':
          description: Referral code applied successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  discount:
                    type: number
                  expiresAt:
                    type: string
                    format: date-time
        '400':
          description: Invalid code or already used
  /referral/available-credits:
    get:
      tags: [Referrals]
      summary: Get user's available credits
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Credits retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  credits:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserCredit'
                  totalAmount:
                    type: string

  # ============ PROMOTIONS ============
  /promotions/banners:
    get:
      tags: [Promotions]
      summary: Get active promotional banners
      description: Public endpoint to get currently active banners
      responses:
        '200':
          description: Banners retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PromotionalBanner'

    post:
      tags: [Promotions]
      summary: Create promotional banner
      description: Admin only
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, imageUrl, startDate, endDate]
              properties:
                title:
                  type: string
                subtitle:
                  type: string
                imageUrl:
                  type: string
                  format: uri
                targetUrl:
                  type: string
                targetScreen:
                  type: string
                priority:
                  type: integer
                isActive:
                  type: boolean
                startDate:
                  type: string
                  format: date-time
                endDate:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Banner created
        '400':
          description: Validation error
        '403':
          description: Admin only
  /promotions/banners/all:
    get:
      tags: [Promotions]
      summary: Get all promotional banners
      description: Admin only - includes inactive banners
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All banners retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PromotionalBanner'
        '403':
          description: Admin only
  /promotions/banners/{id}:
    put:
      tags: [Promotions]
      summary: Update promotional banner
      description: Admin only
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                subtitle:
                  type: string
                imageUrl:
                  type: string
                targetUrl:
                  type: string
                priority:
                  type: integer
                isActive:
                  type: boolean
                startDate:
                  type: string
                  format: date-time
                endDate:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Banner updated
        '404':
          description: Banner not found
        '403':
          description: Admin only

    delete:
      tags: [Promotions]
      summary: Delete promotional banner
      description: Admin only
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Banner deleted
        '404':
          description: Banner not found
        '403':
          description: Admin only

  # ============ NOTIFICATION SETTINGS ============
...