# backend/routes/referrals.js
const express = require('express');
const router = express.Router();
const ReferralCode = require('../models/ReferralCode');
const Referral = require('../models/Referral');
const UserCredit = require('../models/UserCredit');
const User = require('../models/User');
const { protect } = require('../middleware/auth');

// Helper function to generate referral code
const generateReferralCode = async () => {
    // Characters to use: uppercase letters and numbers (excluding similar-looking ones)
    // Excluded: 0, O, I, 1, L to avoid confusion
    const chars = 'ABCDEFGHJKMNPQRSTUVWXYZ23456789';

    // Try to generate a unique 8-character code
    for (let attempt = 0; attempt < 20; attempt++) {
        let code = '';
        for (let i = 0; i < 8; i++) {
            code += chars.charAt(Math.floor(Math.random() * chars.length));
        }

        // Check if code already exists
        const exists = await ReferralCode.findOne({ code });
        if (!exists) {
            return code;
        }
    }

    // Fallback: use timestamp-based code (very unlikely to reach here)
    const timestamp = Date.now().toString(36).toUpperCase();
    const randomSuffix = Math.random().toString(36).substring(2, 6).toUpperCase();
    return `${timestamp}${randomSuffix}`.substring(0, 8);
};

// @route   GET /api/referral/my-code
// @desc    Get user's referral code and stats
// @access  Private
router.get('/my-code', protect, async (req, res) => {
    try {
        let referralCode = await ReferralCode.findOne({ user: req.user._id });

        // Generate code if doesn't exist
        if (!referralCode) {
            const code = await generateReferralCode();

            referralCode = await ReferralCode.create({
                user: req.user._id,
                code
            });
        }

        // Get referral stats
        const totalReferrals = await Referral.countDocuments({ referrer: req.user._id });
        const completedReferrals = await Referral.countDocuments({
            referrer: req.user._id,
            status: 'completed'
        });
        const pendingReferrals = await Referral.countDocuments({
            referrer: req.user._id,
            status: 'pending_order'
        });

        // Get available credits
        const credits = await UserCredit.find({
            user: req.user._id,
            isActive: true,
            usedAt: null,
            expiresAt: { $gt: new Date() }
        });
        const availableCredit = credits.reduce((sum, credit) => sum + credit.amount, 0);

        res.json({
            code: referralCode.code,
            shareUrl: `${process.env.APP_URL || 'https://grabgo.app'}/r/${referralCode.code}`,
            totalReferrals,
            completedReferrals,
            pendingReferrals,
            totalEarned: referralCode.totalEarned,
            availableCredit: availableCredit.toFixed(2)
        });
    } catch (error) {
        console.error('Error getting referral code:', error);
        res.status(500).json({ message: 'Server error' });
    }
});

// @route   GET /api/referral/my-referrals
// @desc    Get list of user's referrals
// @access  Private
router.get('/my-referrals', protect, async (req, res) => {
    try {
        const referrals = await Referral.find({ referrer: req.user._id })
            .populate('referee', 'username')
            .sort({ createdAt: -1 })
            .limit(50);

        const formattedReferrals = referrals.map(ref => ({
            id: ref._id,
            refereeName: ref.referee ? ref.referee.username : 'User',
            status: ref.status,
            creditEarned: ref.status === 'completed' ? ref.referrerCreditAmount : 0,
            createdAt: ref.createdAt,
            completedAt: ref.completedAt
        }));

        res.json({ referrals: formattedReferrals });
    } catch (error) {
        console.error('Error getting referrals:', error);
        res.status(500).json({ message: 'Server error' });
    }
});

// @route   POST /api/referral/validate
// @desc    Validate a referral code
// @access  Public
router.post('/validate', async (req, res) => {
    try {
        const { code } = req.body;

        if (!code) {
            return res.status(400).json({ message: 'Referral code is required' });
        }

        const referralCode = await ReferralCode.findOne({
            code: code.toUpperCase(),
            isActive: true
        });

        if (!referralCode) {
            return res.json({
                valid: false,
                message: 'Invalid referral code'
            });
        }

        // Only populate user for non-system codes
        if (!referralCode.isSystemCode) {
            await referralCode.populate('user', 'username');
        }

        res.json({
            valid: true,
            referrerName: referralCode.isSystemCode ? 'GrabGo' : (referralCode.user?.username || 'Unknown'),
            discount: referralCode.discount || 10.00,
            minOrderValue: referralCode.minOrderValue || 20.00,
            validDays: referralCode.validDays || 7,
            isSystemCode: referralCode.isSystemCode || false
        });
    } catch (error) {
        console.error('Error validating referral code:', error);
        res.status(500).json({ message: 'Server error' });
    }
});

// @route   POST /api/referral/apply
// @desc    Apply referral code to new user
// @access  Private
router.post('/apply', protect, async (req, res) => {
    try {
        const { code } = req.body;

        if (!code) {
            return res.status(400).json({ message: 'Referral code is required' });
        }

        // Check if user already has a referral
        const existingReferral = await Referral.findOne({ referee: req.user._id });
        if (existingReferral) {
            return res.status(400).json({ message: 'You have already used a referral code' });
        }

        // Find referral code
        const referralCode = await ReferralCode.findOne({
            code: code.toUpperCase(),
            isActive: true
        });

        if (!referralCode) {
            return res.status(400).json({ message: 'Invalid referral code' });
        }

        // Can't refer yourself
        if (referralCode.user.toString() === req.user._id.toString()) {
            return res.status(400).json({ message: 'You cannot use your own referral code' });
        }

        // Create referral record FIRST (without credit ID)
        const referralExpiry = new Date();
        referralExpiry.setDate(referralExpiry.getDate() + 7);

        const referral = await Referral.create({
            referrer: referralCode.user,
            referee: req.user._id,
            referralCode: code.toUpperCase(),
            status: 'pending_order',
            expiresAt: referralExpiry,
            deviceId: req.body.deviceId || null,
            ipAddress: req.ip || null
        });

        // Create referee credit with referral link
        const refereeCreditExpiry = new Date();
        refereeCreditExpiry.setDate(refereeCreditExpiry.getDate() + 7);

        const refereeCredit = await UserCredit.create({
            user: req.user._id,
            amount: 10.00,
            source: 'referral_received',
            referralId: referral._id, // ✅ Now we have the referral ID
            expiresAt: refereeCreditExpiry,
            description: `Referral credit from ${code}`
        });

        // Update referral with credit ID
        await Referral.findByIdAndUpdate(referral._id, {
            refereeCreditId: refereeCredit._id
        });

        // Update referral code stats
        await ReferralCode.findByIdAndUpdate(referralCode._id, {
            $inc: { totalReferrals: 1 }
        });

        res.json({
            success: true,
            message: 'Referral code applied! You have GHS 10 off your first order.',
            discount: 10.00,
            expiresAt: refereeCreditExpiry
        });
    } catch (error) {
        console.error('Error applying referral code:', error);
        res.status(500).json({ message: 'Server error' });
    }
});

// @route   GET /api/referral/available-credits
// @desc    Get user's available credits
// @access  Private
router.get('/available-credits', protect, async (req, res) => {
    try {
        const credits = await UserCredit.find({
            user: req.user._id,
            isActive: true,
            usedAt: null,
            expiresAt: { $gt: new Date() }
        }).sort({ expiresAt: 1 }); // Oldest first (FIFO)

        const totalAmount = credits.reduce((sum, credit) => sum + credit.amount, 0);

        res.json({
            credits: credits.map(c => ({
                id: c._id,
                amount: c.amount,
                source: c.source,
                expiresAt: c.expiresAt,
                description: c.description
            })),
            totalAmount: totalAmount.toFixed(2)
        });
    } catch (error) {
        console.error('Error getting credits:', error);
        res.status(500).json({ message: 'Server error' });
    }
});

module.exports = router;